Update UploadDocumentMutation, DeleteDocumentMutation, remove rest api, refactor Query

+++ app/grafql_api/mutations/document.py
@@ -1,6 +1,8 @@
 import graphene
-from django.contrib.auth import get_user_model
 from graphene_file_upload.scalars import Upload
+
+from django.contrib.auth import get_user_model
+
 from graphql_jwt.decorators import login_required
 
 from app.grafql_api.types import DocumentType
@@ -14,16 +16,21 @@ class UploadDocumentMutation(graphene.Mutation):
     class Arguments:
         file = Upload(required=True)
 
-    document_id = graphene.ID()
+    document = graphene.Field(DocumentType)
 
+    @classmethod
     @login_required
-    def mutate(self, info, file, **kwargs):
+    def mutate(cls, root, info, file, **kwargs):
         file_name = file.__dict__["_name"]
-        if Documents.objects.filter(title=file_name).exists():
-            raise Exception("Document already exists")
+        if not cls.check_document_extension(file_name):
+            raise Exception("Document should be in pdf or md format")
         user = User.objects.get(pk=info.context.user.pk)
         document = Documents.objects.create(document=file, title=file_name, owner=user)
-        return UploadDocumentMutation(document_id=document.pk)
+        return UploadDocumentMutation(document=document)
+
+    @staticmethod
+    def check_document_extension(file_name: str) -> bool:
+        return file_name.lower().endswith((".pdf", ".md"))
 
 
 class DeleteDocumentMutation(graphene.Mutation):
@@ -36,4 +43,6 @@ class DeleteDocumentMutation(graphene.Mutation):
     @login_required
     def mutate(cls, root, info, id):
         document = Documents.objects.get(pk=id)
+        if not document.owner.pk == info.context.user.pk:
+            raise Exception("You don't have permission to delete this document")
         document.delete()
